Move models to a separate module
