+++ x10/models/funding_rate.py
@@ -2,7 +2,7 @@ from decimal import Decimal
 
 from pydantic import AliasChoices, Field
 
-from x10.utils.model import X10BaseModel
+from x10.models.base import X10BaseModel
 
 
 class FundingRateModel(X10BaseModel):
